requirements.txt
google-api-python-client==2.54.0
google-auth-oauthlib==0.5.2
python-dateutil==2.8.2
ratelimit==2.2.1
requests==2.22.0
requests-html==0.10.0
titlecase==0.12.0
